requirements.txt
django==4.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pillow==9.2.0
drf-yasg==1.21.3
python-decouple==3.6
gunicorn==20.1.0
whitenoise==6.2.0
dj-database-url==1.0.0
django-cors-headers==3.13.0
